call_center_ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
twilio==8.10.0
openai==1.3.5
anthropic==0.7.7
pydub==0.25.1
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
elevenlabs==0.2.27
google-cloud-texttospeech==2.14.2
SpeechRecognition==3.10.0
gtts==2.4.0
